first successful run
