+++ configs/lra/ds_train_transformer_pathfinder32_layers.sh
